build(front): refactors Code and Link file locations
